refactor Remark Florian
