+++ samples/resilience-examples/src/test-util.ts
@@ -1,9 +1,5 @@
 import { setTestDestination } from '@sap-cloud-sdk/test-util';
 import nock from 'nock';
-import {
-  BusinessPartner,
-  businessPartnerService
-} from '@sap/cloud-sdk-vdm-business-partner-service';
 
 export const destinationName = 'MyDestination';
 export const destinationUrl = 'http://some.url.com';
@@ -20,14 +16,4 @@ export const sampleResponse = {
 
 export function mockGetAllRequest(times: number): void {
   nock(destinationUrl).get(/.*/).times(times).reply(200, sampleResponse);
-}
-
-export async function getAllBusinessPartner(
-  top: number
-): Promise<BusinessPartner[]> {
-  return businessPartnerService()
-    .businessPartnerApi.requestBuilder()
-    .getAll()
-    .top(top)
-    .execute({ destinationName });
 }
